refactor: Update domain alias and rename NavigationButton to WelcomeBlock
