+++ src/domains/Theme/style.css
@@ -1,28 +1,29 @@
 :root {
     --color-text-primary: #ffffff;
-    --primary-fill-color: #FFFFFF;
-    --elements-fill-color: #EFEFEF;
+    --primary-font-color: #111213;
+    --text-inactive: #d2d2d2;
+    --secondary-components-color: #8E8E93;
+    --additional-text-font-color: #678259;
+    
     --primary-background-color: #F4F4F4;
     --secondary-background-color: #3b3a3a;
-    --home-page-background: linear-gradient(#242424, #3b3a3a);
-    --inactive-background-color: #060606;
-    --secondary-fill-color: #727173;
-    --primary-font-color: #111213;
-    --accent-passive-color: #000e8a;
-    --accent-active-color: #2c3e50;
-    --image-background-color: #D9D9D9;
-    --additional-text-font-color: #678259;
-    --secondary-components-color: #8E8E93;
-    --text-inactive: #d2d2d2;
-
     --dark-background: #202020;
+    --inactive-background-color: #060606;
+    --home-page-background: linear-gradient(#242424, #3b3a3a);
     --chat-header-background: #1c1c1d;
     --chat-message-me: #313131;
     --chat-message-other: #262628;
-    --chat-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
+    --image-background-color: #D9D9D9;
+    
+    --primary-fill-color: #FFFFFF;
+    --elements-fill-color: #EFEFEF;
+    --secondary-fill-color: #727173;
     --input-border-color: #3a3a3c;
-    --placeholder-color: #636366;
     --send-button-color: #d500c3;
+    --placeholder-color: #636366;
+
+    --accent-passive-color: #000e8a;
+    --accent-active-color: #2c3e50;
 
     --navbar-height: 75px;
     --product-padding: 10px;
@@ -31,18 +32,20 @@
     --content-radius-secondary: 8px;
     --content-width-desktop: 1100px;
     --control-elements-height: 45px;
-
-    --transition-time: .3s;
-    --disabled-filter: grayscale(.3);
-    --elements-hover-rotation: rotate(4deg);
+    
+    --chat-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
     --default-shadow-color: rgba(23, 28, 23, 0.58);
     --default-shadow: 0 0 4px 0 var(--default-shadow-color);
     --home-intro-slide-shadow: 0 2.02px 6.17px rgb(0, 0, 0, 0.04),
-        0 4.85px 14.83px rgb(0, 0, 0, 0.06),
-        0 9.14px 27.92px rgba(13, 11, 11, 0.07),
-        0 16.31px 49.81px rgb(0, 0, 0, 0.08),
-        0 30.5px 93.16px rgb(0, 0, 0, 0.1),
-        0 73px 223px rgb(0, 0, 0, 0.14);
+    0 4.85px 14.83px rgb(0, 0, 0, 0.06),
+    0 9.14px 27.92px rgba(13, 11, 11, 0.07),
+    0 16.31px 49.81px rgb(0, 0, 0, 0.08),
+    0 30.5px 93.16px rgb(0, 0, 0, 0.1),
+    0 73px 223px rgb(0, 0, 0, 0.14);
+    
+    --transition-time: .3s;
+    --disabled-filter: grayscale(.3);
+    --elements-hover-rotation: rotate(4deg);
 }
 
 @font-face {
@@ -103,6 +106,11 @@ h5 {
     color: var(--primary-font-color);
 }
 
+.sf-pro-text-regular {
+    font-family: "SF Pro Text";
+    font-weight: 400;
+}
+
 .common-text {
     font-weight: 500;
     font-size: 14px;
@@ -127,9 +135,4 @@ h5 {
     font-weight: 700;
     font-size: 16px;
     color: var(--accent-passive-color);
-}
-
-.sf-pro-text-regular {
-    font-family: "SF Pro Text";
-    font-weight: 400;
 }
