+++ tsconfig.app.json
@@ -30,7 +30,7 @@
             "@pages/*": ["src/pages/*"],
             "@app-types/*": ["src/types/*"],
             "@router/*":["src/router/*"],
-            "@domain/*": ["src/domain/*"]
+            "@domains/*": ["src/domains/*"]
         }
     },
     "include": ["src"]
